added startup script

+++ .configs/startup.sh
@@ -1,19 +1,33 @@
 #!/bin/bash
+
+# Start up shell-script to restore dev environment on immutable system
+# Author : Brett Nicholas
+
 cd ~/workspace
 
-sudo apt-get install vim 
-
-#git config --global user.name "Brett: ARMDEV"
-#git config --global user.email Brett.r.nicholas.15@dartmouth.edu
+# ensure git is up to date
 git status
 
-read -p "Would you like to pull from remote git repository? (y/n) > \n" -n 1 
+# Prompt user to pull from remote git repo
+read -p "Would you like to pull from remote git repository? (y/n) > " -n 1 
 echo
 if [[ $REPLY =~ ^[Yy]$ ]]; then 
 	git pull origin master 
 fi 
 
-[! -f ~/workspace/.configs/.vimrc] && [echo ".vimrc not found"; exit 2]
-cat ~/workspace/.configs/.vimrc > ~/.vimrc
-	
-
+# Install Vim because Vi sucks
+echo
+echo "Checking for Vim"
+echo
+sudo apt-get install vim 
+echo
+echo "Checking for .vimrc file..."
+echo
+if test -f "/home/armdev/workspace/.configs/.vimrc"; then {
+	echo
+	echo "updating .vimrc file..."
+	echo
+	cat ~/workspace/.configs/.vimrc > ~/.vimrc
+}; else
+	echo "ERROR... Could not locate ~/workspace/.configs/.vimrc"
+fi
